requirements.txt
openai
python-dotenv
langchain
langchain-openai
psycopg2==2.9.9
pgvector==0.2.5
SQLModel==0.0.16
asyncpg==0.29.0
numpy
pandas
streamlit
streamlit-modal
SQLAlchemy
psycopg2-binary